added air-attacks and added the html file
The html file was added because opening the .sb3 file in the turbowarp engine can cause much lag, the html variant doesn´t take up that many resources. The html file does only represent the current state of the project and will be updated as we update the .sb3 file.